pyproject.toml
[project]
name = "siclo-scraping"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "openai>=2.3.0",
    "pytest-playwright>=0.7.1",
    "selectolax>=0.4.0",
]
